Add unit tests and reorganize some folder structure
