+++ Framework/XOFF.Core/Repository/AutofacRepositoryServiceLocator.cs
