--- Framework/XOFF.Core/SyncRepositorySettings.cs
@@ -1,15 +1,0 @@
-﻿namespace XOFF.Core.Settings
-{
-    public class SyncRepositorySettings
-    {
-        public SyncRepositorySettings(int refreshSeconds = 86400, RefreshDataMode refreshDataMode = Settings.RefreshDataMode.RefreshIfStale)//default to 1 day
-        {
-            RefreshSecondsThreshold = refreshSeconds;
-            RefreshDataMode = refreshDataMode;
-        }
-
-        public RefreshDataMode RefreshDataMode { get; }
-        public int RefreshSecondsThreshold { get; }
-
-    }
-}
