--- XOFF.Tests/XOFF.Tests.Framework/Test.cs
@@ -1,14 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-namespace XOFF.Tests.Framework
-{
-	[TestFixture()]
-	public class Test
-	{
-		[Test()]
-		public void TestCase()
-		{
-			Assert.True(true);
-		}
-	}
-}
