+++ XOFF.Tests/XOFF.Tests.Framework/XOFF.Tests.Framework.csproj
@@ -28,11 +28,17 @@
     <Reference Include="Autofac, Version=4.5.0.0, Culture=neutral, PublicKeyToken=17863af14b0044da, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Autofac.4.5.0\lib\net45\Autofac.dll</HintPath>
     </Reference>
+    <Reference Include="Castle.Core, Version=4.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Castle.Core.4.0.0\lib\net45\Castle.Core.dll</HintPath>
+    </Reference>
     <Reference Include="DBreeze, Version=1.84.2017.321, Culture=neutral, PublicKeyToken=174b57f1c155f3c6, processorArchitecture=MSIL">
       <HintPath>..\..\packages\DBreeze.1.84.0\lib\net45\DBreeze.dll</HintPath>
     </Reference>
     <Reference Include="LiteDB, Version=3.1.0.0, Culture=neutral, PublicKeyToken=4ee40123013c9f27, processorArchitecture=MSIL">
       <HintPath>..\..\packages\LiteDB.3.1.0\lib\net35\LiteDB.dll</HintPath>
+    </Reference>
+    <Reference Include="Moq, Version=4.7.8.0, Culture=neutral, PublicKeyToken=69f491c39445e920, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Moq.4.7.8\lib\net45\Moq.dll</HintPath>
     </Reference>
     <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Newtonsoft.Json.6.0.8\lib\net45\Newtonsoft.Json.dll</HintPath>
@@ -57,11 +63,13 @@
     <Reference Include="System.Net.Http" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Test.cs" />
+    <Compile Include="Core\Repositories\SyncedRepositoryTests.cs" />
+    <Compile Include="Person.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="..\..\Framework\XOFF.SQLite\XOFF.SQLite.projitems" Label="Shared" Condition="Exists('..\..\Framework\XOFF.SQLite\XOFF.SQLite.projitems')" />
   <Import Project="..\..\Framework\XOFF.LiteDB\XOFF.LiteDB.projitems" Label="Shared" Condition="Exists('..\..\Framework\XOFF.LiteDB\XOFF.LiteDB.projitems')" />
   <Import Project="..\..\Framework\XOFF.Core\XOFF.Core.projitems" Label="Shared" Condition="Exists('..\..\Framework\XOFF.Core\XOFF.Core.projitems')" />
