+++ XOFF.Tests/XOFF.Tests.Framework/packages.config
@@ -1,8 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Autofac" version="4.5.0" targetFramework="net45" />
+  <package id="Castle.Core" version="4.0.0" targetFramework="net45" />
   <package id="DBreeze" version="1.84.0" targetFramework="net45" />
   <package id="LiteDB" version="3.1.0" targetFramework="net45" />
+  <package id="Moq" version="4.7.8" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
   <package id="SQLite.Net.Core-PCL" version="3.1.1" targetFramework="net45" />
